requirements.txt
certifi==2018.8.24
chardet==3.0.4
hvac==0.6.4
idna==2.7
PyYAML==3.13
requests==2.19.1
urllib3==1.23
